Adds project resource & http error handling

+++ app/Http/Controllers/Api/ProjectUserController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers\Api;
 
+use App\Models\User;
 use App\Models\Project;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
+use Symfony\Component\HttpFoundation\Response as HttpResponse;
 
 class ProjectUserController extends Controller
 {
@@ -18,10 +20,24 @@ class ProjectUserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function show($projectId, $userId) {
-        return DB::table('project_user')
+        if(DB::table('project_user')
+                ->where('project_id', $projectId)
+                ->where('user_id', $userId)
+                ->doesntExist())
+        {
+            return response()->json([
+                'error' => 'The user is not part of the project'
+            ], $status = HttpResponse::HTTP_NOT_FOUND);
+        }
+
+        $projectUser = DB::table('project_user')
                 ->where('project_id', $projectId)
                 ->where('user_id', $userId)
                 ->get();
+
+        return response()->json([
+            'data' => $projectUser
+        ], $status = HttpResponse::HTTP_OK);
     }
 
     /**
@@ -36,7 +52,9 @@ class ProjectUserController extends Controller
         $project = Project::find($projectId);
         $project->users()->attach($userId, array('project_score' => 0));
 
-        return response(200);
+        return response()->json([
+            'data' => $project
+        ], $status = HttpResponse::HTTP_CREATED);
     }
 
     /**
@@ -49,6 +67,16 @@ class ProjectUserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update($projectId, $userId, Request $request) {
+        if(DB::table('project_user')
+                ->where('project_id', $projectId)
+                ->where('user_id', $userId)
+                ->doesntExist())
+        {
+            return response()->json([
+                'error' => 'The user is not part of the project'
+            ], $status = HttpResponse::HTTP_NOT_FOUND);
+        }
+
         DB::table('project_user')
             ->where('project_id', $projectId)
             ->where('user_id', $userId)
@@ -56,7 +84,7 @@ class ProjectUserController extends Controller
                 array("project_score" => $request->input('project_score'))
             );
 
-        return response(200);
+        return response()->noContent();
 
     }
 
@@ -69,15 +97,17 @@ class ProjectUserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($projectId, $userId) {
-        // DB::table('project_user')
-        //     ->where('project_id', $projectId)
-        //     ->where('user_id', $userId)
-        //     ->delete();
-
         $project = Project::find($projectId);
+        $user = User::find($userId);
+        if(!$project || !$user) {
+            return response()->json([
+                'error' => 'The user is not part of the project'
+            ], $status = HttpResponse::HTTP_NOT_FOUND);
+        }
+
         $project->users()->detach($userId);
 
-        return response(200);
+        return response()->noContent();
 
     }
